+++ lib/presentation/pages/home/core/error_page.dart
